Update README.md
Fetch data from the Google Books API and show an initial book on the first page. Implement an Intersection Observer to seamlessly load additional book suggestions when reaching the footer, enhancing the user experience. Apply a skeleton effect during the loading.When clicking one of the suggestions, Navigate to the second page to explore in-depth details about the selected book. Users can add items to the cart and mark them as favourites, utilizing local storage for data persistence.